layers: Move External object CoreCheck into own file
